requirements.txt
pyyaml==3.12
requests==2.18.4
boto3==1.4.7
Flask==0.12.2
Flask-SQLAlchemy==2.2
Flask-Migrate==2.1.1
flask-sslify==0.1.5
flask_restplus==0.10.1
flask-script==2.0.5
psycopg2==2.7.3.1
apscheduler==3.3.1